+++ src/exporters/aws/AWSS3SpanExporter.js
